Separate out PolysemyStuff
